bug fixes on the ethernet shield, including reducing socket num
The ethernet shield was originally configured as a generic component
with 8 sockets, but because we are using a raw socket, we can change
this to using only 1 socket. This allowed us to expand the size of
the internal buffers to 2048 instead of the VERY limiting 256. There
were some operator precedence bugs fixed here as well.

Next step is to fix the state machine so that transfers of over 255 bytes
use multiple steps -- this is due to the constraint of the SPI interface
that only allows uint8_t as the length parameter.