Use uint16_t lengths to transfer SPI packets > 255 bytes
Turns out the W5200 supports 15-bit lengths on packets, which is more than enough
to cover the MTU of the links we'll be using. This commit adjusts the interface
in SocketSpi and SocketSpiP implementation to use longer lengths in order to avoid
the trncation we had before. I'm currently observing the border router handle 1kB
payload pings for over 5 minutes