Load forwarding prefix from FlashAttr
Like before, this will only use the prefix in attribute 2, and will only
use it if it is actually filled (length > 0), or it will default to
using IN6_PREFIX from the Makefile. This also wires up the
IPForwardingP.Init method to MainC.SoftwareInit because that wasn't
being called before. It memsets the whole routing table to 0 on startup,
which I figure can't be a bad thing to do